Added test for second implementation
