+++ CanHazFunny/CanHazFunny.Tests/JokeServiceTests.cs
@@ -14,7 +14,7 @@ public class JokeServiceTests
     public void GetJoke_NoArgs_ReturnsJoke()
     {
         // Arrange
-        JsonJokeService service = new();
+        JokeService service = new();
         string joke;
 
         // Act
@@ -23,23 +23,5 @@ public class JokeServiceTests
         // Assert
         Assert.NotNull(joke);
         Assert.NotEqual("", joke);
-        //Assert.AreEqual("hi", joke); //Uncomment line to see result
-    }
-
-
-    [Theory]
-    [InlineData("{ \"joke\": \"A joke here!\" }", "A joke here!")]
-    public void ParseJsonJoke_ValidJson_ReturnsString(string jsonString, string expected)
-    {
-        // Arrange
-        JsonJokeService service = new();
-        string? response;
-
-        // Act
-        response =service.ParseJokeFromJsonString(jsonString);
-        // Assert
-        Assert.NotNull(response);
-        Assert.Equal(expected, response);
-        //Assert.AreEqual("expected:"+expected, response); //Uncomment line to see result
     }
 }
